* vm_backtrace.c: use `long' for return values of `NUM2LONG()'.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@37961 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ vm_backtrace.c
@@ -698,7 +698,7 @@ static VALUE
 vm_backtrace_to_ary(rb_thread_t *th, int argc, VALUE *argv, int lev_default, int lev_plus, int to_str)
 {
     VALUE level, vn;
-    int lev, n;
+    long lev, n;
     VALUE btval = backtrace_object(th);
     rb_backtrace_t *bt;
     GetCoreDataFromValue(btval, rb_backtrace_t, bt);
